:truck: api: move CertifiedBadge read model from lib to src
