+++ api/src/certification/results/domain/read-models/CertifiedBadge.js
